+++ src/Raven.CodeAnalysis/Symbols/ISymbol.cs
