+++ src/Raven.CodeAnalysis/Symbols/PE/MetadataAssemblySymbol.cs
