+++ src/Raven.CodeAnalysis/Symbols/PE/MetadataModuleSymbol.cs
