+++ src/Raven.CodeAnalysis/Symbols/PE/MetadataNamedTypeSymbol.cs
