+++ src/Raven.CodeAnalysis/Symbols/PE/MetadataSymbol.cs
