+++ src/Raven.CodeAnalysis/Symbols/Source/SourceAssemblySymbol.cs
