+++ src/Raven.CodeAnalysis/Symbols/Source/SourceNamedTypeSymbol.cs
